finished authentication feature

+++ src/project/Pickler.java
@@ -1,10 +1,6 @@
 package project;
 
-import java.util.List;
-
 import javax.crypto.Cipher;
-import javax.crypto.KeyGenerator;
-import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
 
 import java.io.FileOutputStream;
@@ -52,7 +48,7 @@ public class Pickler {
         // Encrypt file f using AES
         try {
             fileProcessor(Cipher.ENCRYPT_MODE,f,encryptedFile);
-            System.out.println("Success");
+            System.out.println("Encrypt success");
         } catch (Exception ex) {
             System.out.println(ex.getMessage());
             ex.printStackTrace();
@@ -66,32 +62,44 @@ public class Pickler {
         File f = new File(filename);
         File encryptedFile = new File("text.encrypted");
 
-        // Decrypt file
-        try {
-            f.createNewFile();
-            fileProcessor(Cipher.DECRYPT_MODE,encryptedFile,f);
-            System.out.println("Success");
-        } catch (Exception ex) {
-            System.out.println(ex.getMessage());
-                ex.printStackTrace();
-        }
+        if (encryptedFile.exists() && !encryptedFile.isDirectory()) {
+            // Decrypt file
+            try {
+                f.createNewFile();
+                fileProcessor(Cipher.DECRYPT_MODE,encryptedFile,f);
+                System.out.println("Decrypt success");
+            } catch (Exception ex) {
+                System.out.println(ex.getMessage());
+                    ex.printStackTrace();
+            }
 
-        // Load passwords into manager
-        try {
-            FileInputStream fileIn = new FileInputStream(f);
-            ObjectInputStream objIn = new ObjectInputStream(fileIn);
-            manager = (Manager) objIn.readObject();
-            fileIn.close();
-            objIn.close();
-        } catch (IOException e) {
-            System.out.println("A file load error has occured");
-            e.printStackTrace();
-        } catch (ClassNotFoundException e) {
-            System.out.println("A class error has occured");
-            e.printStackTrace();
+            // Load passwords into manager
+            try {
+                FileInputStream fileIn = new FileInputStream(f);
+                ObjectInputStream objIn = new ObjectInputStream(fileIn);
+                manager = (Manager) objIn.readObject();
+                fileIn.close();
+                objIn.close();
+            } catch (IOException e) {
+                System.out.println("A file load error has occured");
+                e.printStackTrace();
+            } catch (ClassNotFoundException e) {
+                System.out.println("A class error has occured");
+                e.printStackTrace();
+            }
+            f.delete();
+            System.out.println("Passwords loaded");
+        } else {
+            System.out.println("Creating new password profile");
+            try {
+                encryptedFile.createNewFile();
+            } catch (IOException e) {
+                System.out.println("Error creating new file");
+                e.printStackTrace();
+            }
+            manager = new Manager();
         }
-        f.delete();
-        System.out.println("Passwords loaded");
+        
         return manager;
     }
